Rough prototype of network placement MCE #122
This is a rough prototype for the MCE dedicated to network placement.

It is designed in such a way that the user just needs to specify the
name under the spa ontology. The system will then proceed to create an
appropriate delta containing all the model information for that network

It has not been tested yet, Openstack compatibility logic needs to be
added