big fix in routing info part
right now it support the multiple routing info create at one time. But
not tested it yet, will generate some delta to test, the test based on
one router is very successful. Using some new data structure to import
the routing info and to get them in the commit part, not only fix the
problem but also enhance the speed