Functional network placement MCE #122
The MCE for network placement is functional now where a network can be
created with multiple subnets and gateways. There is some more logic
too be designed for certain special cases and for IP CIDR block check.

Need to fix a typo on the exportPolicyData method that prevents the
newly created resources from appearing in the spa model as policy data

Fixed bug in AWS push that prevented from catching an exception while
creating subnet with no IP address