Added new MCE for virtual cloud creation, fixed bug in AWS push
Added new MCE for virtual cloud creation to create networks with
subnets, gateways and routes as explained in issue #122 Needs a little
more development and testing. Initial testing passes for network and
gateway creations with route addition to the main table. Subnet routing
is still in development

Fixed a bug in the AWS push that prevented the creation of a gateway
when new tags were specified in a delta model. This bug was affecting
the creation of a gateway through the MCE, bug was fixed and verified