Functional MCE for virtual cloud creation #122
Functional MCE for virtual cloud creations. This MCE is capable of
adding networks, subnets, gateways,routing tables and routes when a
JSON spa model with an import statement composed of a JSON string is
entered.

This MCE was tested against the AWS driver. Testing on OpenStack is
needed

Added VPN propagation capabilities to AWS push so that vpn gateways can
propagate routes

Fixed issue that prevented the differentiation between vpn propagation
and route creation