ModelBuilder now reads vpn endpoint URIs from server metadata key "ipsec:uri". If the uri field is empty, the vpn endpoint URI will default to the VM uri + ":vpn"
Also, defaults for the PSKs have been provided. Thus, server metadata for ipsec.secret_n
will be "PSK_n" where n is the number of the tunnel. This allows strongswan setup to run
on a vm without error even if PSKs have not been provided, although the keys would still
have to be entered manually later.